Api class refactored
